[FIX] AG-3 getState: cast T->float so the double specialisation compiles
Limiter::getState (ceiling/release/lookahead), NoiseGate::getState (threshold/
hysteresis/attack/hold/release/range/scHpfFreq) and Gain::getState (gainLinear)
passed an unqualified T load to StateWriter::write(). For the supported
T=double this is an ambiguous overload (float/int32/bool), so the public
getState() failed to COMPILE for double - a latent break of a supported type
(Compressor/Expander/etc. already cast to float, and the blob format stores
float32). Cast each numeric field to float, matching the rest of the framework.

Pin: State_roundtrip_dynamics_double instantiates getState()/setState() for
T=double across every AG-3 dynamics/gain header and round-trips them; reverting
any cast restores the ambiguous-overload compile error.

+++ Effects/Gain.h
@@ -264,7 +264,7 @@ public:
     [[nodiscard]] std::vector<uint8_t> getState() const
     {
         StateWriter w(stateId("GAIN"), 1);
-        w.write("gainLinear", targetGainLinear_.load(std::memory_order_relaxed));
+        w.write("gainLinear", static_cast<float>(targetGainLinear_.load(std::memory_order_relaxed)));
         w.write("rampMs", static_cast<float>(rampTimeMs_.load(std::memory_order_relaxed)));
         w.write("muted", muted_.load(std::memory_order_relaxed));
         w.write("inverted", inverted_.load(std::memory_order_relaxed));
